// servers/src/types/room.type.ts
export interface UpdateRoom {
  name?: string;
  price?: number;
  stock?: number;
  categoryId?: number;
  countUser?: number;
  linkImage1?: string;
  linkImage2?: string;
  linkImage3?: string;
  linkImage4?: string;
  linkImage5?: string;
  id?: number;
  isDelete?: number;
  createdAt?: string;
  updatedAt?: string;
}
export interface CreateRoomType {
  name?: string;
  price?: number;
  stock?: number;
  categoryId?: number;
  countUser?: number;
  linkImage1?: string;
  linkImage2?: string;
  linkImage3?: string;
  linkImage4?: string;
  linkImage5?: string;
}